+++ lib/preference_page.dart
@@ -1,6 +1,7 @@
 import 'package:flutter/material.dart';
 import 'package:preferences/preference_service.dart';
 
+/// PreferencePage isn't required if you init PrefService in your main() function
 class PreferencePage extends StatefulWidget {
   final List preferences;
   PreferencePage(this.preferences);